Move terms  parallel collection tests to TermsTests
They were before in AggregatorFactoriesTests, but TermsTests is a more appropriate place for them.
