+++ server/src/test/java/org/elasticsearch/search/aggregations/AggregatorFactoriesTests.java
@@ -344,13 +344,6 @@ public class AggregatorFactoriesTests extends ESTestCase {
         }
         {
             AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            builder.addAggregator(
-                new CompositeAggregationBuilder("composite", Collections.singletonList(new TermsValuesSourceBuilder("name")))
-            );
-            assertTrue(builder.supportsParallelCollection(randomCardinality));
-        }
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
             builder.addAggregator(new FilterAggregationBuilder("terms", new MatchAllQueryBuilder()) {
                 @Override
                 public boolean isInSortOrderExecutionRequired() {
@@ -358,69 +351,6 @@ public class AggregatorFactoriesTests extends ESTestCase {
                 }
             });
             assertFalse(builder.supportsParallelCollection(randomCardinality));
-        }
-    }
-
-    public void testSupportsParallelCollectionTermsAgg() {
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            TermsAggregationBuilder terms = new TermsAggregationBuilder("terms").executionHint("map");
-            builder.addAggregator(terms);
-            assertFalse(builder.supportsParallelCollection(field -> randomIntBetween(-1, 100)));
-        }
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            TermsAggregationBuilder terms = new TermsAggregationBuilder("terms").executionHint("global_ordinals");
-            builder.addAggregator(terms);
-            assertTrue(builder.supportsParallelCollection(field -> 0));
-        }
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            TermsAggregationBuilder terms = new TermsAggregationBuilder("terms").order(BucketOrder.key(randomBoolean()));
-            if (randomBoolean()) {
-                terms.shardSize(randomIntBetween(1, 100));
-            }
-            builder.addAggregator(terms);
-            assertTrue(builder.supportsParallelCollection(field -> randomIntBetween(0, 49)));
-        }
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            TermsAggregationBuilder terms = new TermsAggregationBuilder("terms").order(BucketOrder.key(randomBoolean()));
-            if (randomBoolean()) {
-                terms.shardSize(randomIntBetween(1, 100));
-            }
-            builder.addAggregator(terms);
-            assertFalse(builder.supportsParallelCollection(field -> randomIntBetween(51, 100)));
-        }
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            builder.addAggregator(new TermsAggregationBuilder("terms"));
-            assertFalse(builder.supportsParallelCollection(field -> -1));
-        }
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            builder.addAggregator(new TermsAggregationBuilder("terms"));
-            assertTrue(builder.supportsParallelCollection(field -> 0));
-        }
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            TermsAggregationBuilder terms = new TermsAggregationBuilder("terms");
-            terms.subAggregation(new TermsAggregationBuilder("name") {
-                @Override
-                public boolean supportsParallelCollection(ToLongFunction<String> fieldCardinalityResolver) {
-                    return false;
-                }
-            });
-            builder.addAggregator(terms);
-            assertFalse(builder.supportsParallelCollection(field -> 0));
-        }
-        {
-            AggregatorFactories.Builder builder = new AggregatorFactories.Builder();
-            TermsAggregationBuilder terms = new TermsAggregationBuilder("terms");
-            terms.shardSize(10);
-            builder.addAggregator(terms);
-            assertTrue(builder.supportsParallelCollection(field -> randomIntBetween(1, 10)));
-            assertFalse(builder.supportsParallelCollection(field -> randomIntBetween(11, 100)));
         }
     }
